+++ src/main/java/com/intellij/openapi/fileEditor/impl/ForwardInDocumentAction.java
@@ -1,0 +1,30 @@
+package com.intellij.openapi.fileEditor.impl;
+
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Objects;
+
+public class ForwardInDocumentAction extends AnAction {
+  @Override
+  public void actionPerformed(@NotNull final AnActionEvent e) {
+    getHistory(e).forwardInCurrentDocument();
+  }
+
+  @Override
+  public void update(@NotNull final AnActionEvent e) {
+    super.update(e);
+    e.getPresentation().setEnabled(getHistory(e).getForwardInCurrentDocumentPlace().isPresent());
+  }
+
+  private MyIdeDocumentHistoryImpl getHistory(final AnActionEvent e) {
+    return ((MyIdeDocumentHistoryImpl) MyIdeDocumentHistoryImpl.getInstance(Objects.requireNonNull(e.getProject())));
+  }
+
+  @Override
+  public @NotNull ActionUpdateThread getActionUpdateThread() {
+    return ActionUpdateThread.EDT; // we're accessing current editor
+  }
+}